+++ pkg/reconciler/logging/logging_reconciler.go
@@ -7,6 +7,7 @@ import (
 	"github.com/pkg/errors"
 	apimachineryerrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/runtime"
+	capi "sigs.k8s.io/cluster-api/api/v1beta1"
 	"sigs.k8s.io/cluster-api/util/patch"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -16,7 +17,6 @@ import (
 	"github.com/giantswarm/logging-operator/pkg/common"
 	"github.com/giantswarm/logging-operator/pkg/config"
 	"github.com/giantswarm/logging-operator/pkg/key"
-	loggedcluster "github.com/giantswarm/logging-operator/pkg/logged-cluster"
 	"github.com/giantswarm/logging-operator/pkg/reconciler"
 )
 
@@ -28,39 +28,39 @@ type LoggingReconciler struct {
 	Config      config.Config
 }
 
-func (l *LoggingReconciler) Reconcile(ctx context.Context, lc loggedcluster.Interface) (result ctrl.Result, err error) {
-	if common.IsLoggingEnabled(lc, l.Config.EnableLoggingFlag) {
-		result, err = l.reconcileCreate(ctx, lc)
+func (l *LoggingReconciler) Reconcile(ctx context.Context, cluster *capi.Cluster) (result ctrl.Result, err error) {
+	if common.IsLoggingEnabled(cluster, l.Config.EnableLoggingFlag) {
+		result, err = l.reconcileCreate(ctx, cluster)
 	} else {
-		result, err = l.reconcileDelete(ctx, lc)
+		result, err = l.reconcileDelete(ctx, cluster)
 	}
 
 	return result, errors.WithStack(err)
 }
 
 // reconcileCreate handles creation/update logic by calling ReconcileCreate method on all l.Reconcilers.
-func (l *LoggingReconciler) reconcileCreate(ctx context.Context, lc loggedcluster.Interface) (ctrl.Result, error) {
+func (l *LoggingReconciler) reconcileCreate(ctx context.Context, cluster *capi.Cluster) (ctrl.Result, error) {
 	logger := log.FromContext(ctx)
 	logger.Info("LOGGING enabled")
 
-	if !controllerutil.ContainsFinalizer(lc, key.Finalizer) {
+	if !controllerutil.ContainsFinalizer(cluster, key.Finalizer) {
 		logger.Info("adding finalizer", "finalizer", key.Finalizer)
 
 		// We use a patch rather than an update to avoid conflicts when multiple controllers are adding their finalizer to the ClusterCR
 		// We use the patch from sigs.k8s.io/cluster-api/util/patch to handle the patching without conflicts
-		patchHelper, err := patch.NewHelper(lc.GetObject(), l.Client)
+		patchHelper, err := patch.NewHelper(cluster, l.Client)
 		if err != nil {
 			return ctrl.Result{}, errors.WithStack(err)
 		}
-		controllerutil.AddFinalizer(lc, key.Finalizer)
-		if err := patchHelper.Patch(ctx, lc.GetObject()); err != nil {
+		controllerutil.AddFinalizer(cluster, key.Finalizer)
+		if err := patchHelper.Patch(ctx, cluster); err != nil {
 			logger.Error(err, "failed to add finalizer to logger cluster", "finalizer", key.Finalizer)
 			return ctrl.Result{}, errors.WithStack(err)
 		}
 		logger.Info("successfully added finalizer to logged cluster", "finalizer", key.Finalizer)
 	}
 
-	err := common.ToggleAgents(ctx, l.Client, lc)
+	loggingAgentConfig, err := common.ToggleAgents(ctx, l.Client, cluster, l.Config)
 	if err != nil {
 		// Handle case where the app is not found.
 		if apimachineryerrors.IsNotFound(err) {
@@ -73,7 +73,7 @@ func (l *LoggingReconciler) reconcileCreate(ctx context.Context, lc loggedcluste
 
 	// Call all reconcilers ReconcileCreate methods.
 	for _, reconciler := range l.Reconcilers {
-		result, err := reconciler.ReconcileCreate(ctx, lc)
+		result, err := reconciler.ReconcileCreate(ctx, cluster, loggingAgentConfig)
 		if err != nil || !result.IsZero() {
 			return result, errors.WithStack(err)
 		}
@@ -83,12 +83,13 @@ func (l *LoggingReconciler) reconcileCreate(ctx context.Context, lc loggedcluste
 }
 
 // reconcileDelete handles deletion logic by calling reconcileDelete method on all l.Reconcilers.
-func (l *LoggingReconciler) reconcileDelete(ctx context.Context, lc loggedcluster.Interface) (ctrl.Result, error) {
+func (l *LoggingReconciler) reconcileDelete(ctx context.Context, cluster *capi.Cluster) (ctrl.Result, error) {
 	logger := log.FromContext(ctx)
 	logger.Info("LOGGING disabled")
 
-	if controllerutil.ContainsFinalizer(lc, key.Finalizer) {
-		err := common.ToggleAgents(ctx, l.Client, lc)
+	if controllerutil.ContainsFinalizer(cluster, key.Finalizer) {
+		// Get the current logging agent configuration
+		loggingAgentConfig, err := common.ToggleAgents(ctx, l.Client, cluster, l.Config)
 		if err != nil && !apimachineryerrors.IsNotFound(err) {
 			// Errors only if this is not a 404 because the apps are already deleted.
 			return ctrl.Result{}, errors.WithStack(err)
@@ -96,7 +97,7 @@ func (l *LoggingReconciler) reconcileDelete(ctx context.Context, lc loggedcluste
 
 		// Call all reconcilers ReconcileDelete methods.
 		for _, reconciler := range l.Reconcilers {
-			result, err := reconciler.ReconcileDelete(ctx, lc)
+			result, err := reconciler.ReconcileDelete(ctx, cluster, loggingAgentConfig)
 			if err != nil || !result.IsZero() {
 				return result, errors.WithStack(err)
 			}
@@ -108,12 +109,12 @@ func (l *LoggingReconciler) reconcileDelete(ctx context.Context, lc loggedcluste
 
 		// We use a patch rather than an update to avoid conflicts when multiple controllers are removing their finalizer from the ClusterCR
 		// We use the patch from sigs.k8s.io/cluster-api/util/patch to handle the patching without conflicts
-		patchHelper, err := patch.NewHelper(lc.GetObject(), l.Client)
+		patchHelper, err := patch.NewHelper(cluster, l.Client)
 		if err != nil {
 			return ctrl.Result{}, errors.WithStack(err)
 		}
-		controllerutil.RemoveFinalizer(lc, key.Finalizer)
-		if err := patchHelper.Patch(ctx, lc.GetObject()); err != nil {
+		controllerutil.RemoveFinalizer(cluster, key.Finalizer)
+		if err := patchHelper.Patch(ctx, cluster); err != nil {
 			logger.Error(err, "failed to remove finalizer from logger cluster, requeuing", "finalizer", key.Finalizer)
 			return ctrl.Result{}, errors.WithStack(err)
 		}