+++ pkg/resource/events-logger-secret/reconciler.go
@@ -8,14 +8,13 @@ import (
 	v1 "k8s.io/api/core/v1"
 	apimachineryerrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/types"
-
-	"github.com/giantswarm/logging-operator/pkg/common"
-	loggedcluster "github.com/giantswarm/logging-operator/pkg/logged-cluster"
-	loggingcredentials "github.com/giantswarm/logging-operator/pkg/resource/logging-credentials"
-
+	capi "sigs.k8s.io/cluster-api/api/v1beta1"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/log"
+
+	"github.com/giantswarm/logging-operator/pkg/common"
+	loggingcredentials "github.com/giantswarm/logging-operator/pkg/resource/logging-credentials"
 )
 
 // Reconciler implements a reconciler.Interface to handle
@@ -25,7 +24,7 @@ type Reconciler struct {
 }
 
 // ReconcileCreate ensures events-logger-secret is created with the right credentials
-func (r *Reconciler) ReconcileCreate(ctx context.Context, lc loggedcluster.Interface) (ctrl.Result, error) {
+func (r *Reconciler) ReconcileCreate(ctx context.Context, cluster *capi.Cluster, loggingAgent *common.LoggingAgent) (ctrl.Result, error) {
 	logger := log.FromContext(ctx)
 	logger.Info("events-logger-secret create")
 
@@ -38,13 +37,13 @@ func (r *Reconciler) ReconcileCreate(ctx context.Context, lc loggedcluster.Inter
 	}
 
 	// Retrieve Loki ingress name
-	lokiURL, err := common.ReadLokiIngressURL(ctx, lc, r.Client)
+	lokiURL, err := common.ReadLokiIngressURL(ctx, cluster, r.Client)
 	if err != nil {
 		return ctrl.Result{}, errors.WithStack(err)
 	}
 
 	// Get desired secret
-	desiredEventsLoggerSecret, err := generateEventsLoggerSecret(lc, &eventsLoggerCredentialsSecret, lokiURL)
+	desiredEventsLoggerSecret, err := generateEventsLoggerSecret(cluster, loggingAgent, &eventsLoggerCredentialsSecret, lokiURL)
 	if err != nil {
 		logger.Error(err, "failed generating events logger secret")
 		return ctrl.Result{}, errors.WithStack(err)
@@ -82,13 +81,13 @@ func (r *Reconciler) ReconcileCreate(ctx context.Context, lc loggedcluster.Inter
 }
 
 // ReconcileDelete - Not much to do here when a cluster is deleted
-func (r *Reconciler) ReconcileDelete(ctx context.Context, lc loggedcluster.Interface) (ctrl.Result, error) {
+func (r *Reconciler) ReconcileDelete(ctx context.Context, cluster *capi.Cluster, loggingAgent *common.LoggingAgent) (ctrl.Result, error) {
 	logger := log.FromContext(ctx)
 	logger.Info("events-logger-secret delete")
 
 	// Get expected secret.
 	var currentEventsLoggerSecret v1.Secret
-	err := r.Client.Get(ctx, types.NamespacedName{Name: GetEventsLoggerSecretName(lc), Namespace: lc.GetNamespace()}, &currentEventsLoggerSecret)
+	err := r.Client.Get(ctx, types.NamespacedName{Name: GetEventsLoggerSecretName(cluster, loggingAgent), Namespace: cluster.GetNamespace()}, &currentEventsLoggerSecret)
 	if err != nil {
 		if apimachineryerrors.IsNotFound(err) {
 			logger.Info("events-logger-secret not found, stop here")